created ChromatogramWriter interface made XDecoderException subclass IOException
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@329 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/trace/sanger/chromatogram/ztr/data/DeltaEncodedData.java
@@ -158,7 +158,7 @@ public enum DeltaEncodedData implements Data {
 		}
 		ByteBuffer result = ByteBuffer.allocate(data.length +2+getPaddingSize());
 		result.put(headerByte);
-		result.put((byte)level);
+		result.put(level);
 		for(int i=0; i< getPaddingSize(); i++){
 			result.put((byte)0);
 		}
